+++ system/ping6/ping6.c
@@ -278,7 +278,7 @@ int main(int argc, FAR char *argv[])
   info.callback  = ping6_result;
   info.priv      = &priv;
   priv.code      = ICMPv6_I_OK;
-  priv.tmin      = INT_MAX;
+  priv.tmin      = LONG_MAX;
   priv.tmax      = 0;
   priv.tsum      = 0;
   priv.tsum2     = 0;
